Rename example Activity to FullscreenExampleActivity
