+++ core-sample-app/src/main/AndroidManifest.xml
@@ -86,10 +86,10 @@
             android:hardwareAccelerated="true"
             android:label="@string/simple_example" />
         <activity
-            android:name="com.pierfrancescosoffritti.androidyoutubeplayer.core.sampleapp.examples.simpleFullscreenExample.SimpleFullscreenExampleActivity"
+            android:name="com.pierfrancescosoffritti.androidyoutubeplayer.core.sampleapp.examples.fullscreenExample.FullscreenExampleActivity"
             android:configChanges="orientation|screenSize|keyboardHidden|smallestScreenSize|screenLayout"
             android:hardwareAccelerated="true"
-            android:label="@string/simple_fullscreen_example" />
+            android:label="@string/fullscreen_example" />
         <activity
             android:name="com.pierfrancescosoffritti.androidyoutubeplayer.core.sampleapp.examples.noLifecycleObserverExample.NoLifecycleObserverExampleActivity"
             android:configChanges="orientation|screenSize|keyboardHidden|smallestScreenSize|screenLayout"
