+++ core-sample-app/src/main/res/layout/activity_fullscreen_example.xml
@@ -14,8 +14,8 @@
 
     <FrameLayout
         android:id="@+id/full_screen_view_container"
-        android:layout_width="0dp"
-        android:layout_height="0dp"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
         android:visibility="gone"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
