+++ core-sample-app/src/main/res/values/strings.xml
@@ -1,7 +1,7 @@
 <resources>
     <string name="app_name">ayp API</string>
     <string name="simple_example">Simple example</string>
-    <string name="simple_fullscreen_example">Simple Fullscreen example</string>
+    <string name="fullscreen_example">Fullscreen example</string>
     <string name="complete_example">Complete example</string>
     <string name="recycler_view_example">RecyclerView example</string>
     <string name="custom_ui_example">Custom UI example</string>
